split save messge functionality out into lib/savemessages.js
