+++ index.js
@@ -1,7 +1,4 @@
-var AWS = require('aws-sdk');
-AWS.config.region = 'eu-west-1';
-var bucket = 'dwyl';
-var s3bucket = new AWS.S3({params: {Bucket: bucket}});
+var savemessages = require('./lib/savemessages');
 
 function handler (event, context) {
   console.log('event', JSON.stringify(event));
@@ -13,14 +10,7 @@ function handler (event, context) {
     return context.fail('no message in event');
   }
 
-  var params = {
-    Key: 'chat/' + event.t + '.json',
-    Body: JSON.stringify(event),
-    ContentType: 'application/json',
-    ACL: 'public-read'
-  };
-
-  s3bucket.upload(params, function(err, data) {
+  savemessages(event, function(err, data) {
     if (err) {
       console.log("Error uploading data: ", err);
     } else {
